refactor: 💡 Added default imajsConf

+++ packages/cli/scripts/build.js
@@ -12,7 +12,7 @@ const sharedArgs = require('./lib/sharedArgs');
 async function build({ options, imaConf }) {
   // Clean build directory
   const buildDir = path.join(options.rootDir, 'build');
-  if (fs.existsSync(buildDir)) {
+  if (options.clean && fs.existsSync(buildDir)) {
     fs.rmSync(buildDir, { recursive: true });
   }
 
@@ -30,6 +30,10 @@ const buildCommand = {
     ...sharedArgs,
     compress: {
       desc: 'Compresses resulted assets for use in content-encoding serving',
+      type: 'boolean'
+    },
+    clean: {
+      desc: 'Clean build folder before building the application',
       type: 'boolean'
     }
   },
